back/requirements.txt
pymysql
bcrypt
pyjwt
